+++ src/commonMain/kotlin/io/github/koalaplot/core/xygraph/IntLinearAxisModel.kt
@@ -1,10 +1,9 @@
 package io.github.koalaplot.core.xygraph
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
+import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import io.github.koalaplot.core.util.sign
@@ -65,10 +64,14 @@ public class IntLinearAxisModel(
     }
 
     // Internal for testing
-    internal var currentRange by mutableStateOf(range.first..(range.first + maxViewExtent))
+    internal var currentRange = mutableStateOf(range.first..(range.first + maxViewExtent))
+    public override val viewRange: State<ClosedRange<Int>> = currentRange
 
     override fun computeOffset(point: Int): Float {
-        return ((point - currentRange.first).toFloat() / (currentRange.last - currentRange.first).toFloat())
+        return (
+            (point - currentRange.value.first).toFloat() /
+                (currentRange.value.last - currentRange.value.first).toFloat()
+            )
     }
 
     /**
@@ -82,14 +85,14 @@ public class IntLinearAxisModel(
 
         return buildList {
             if (tickSpacing > 0) {
-                var tickCount = currentRange.first / tickSpacing
+                var tickCount = currentRange.value.first / tickSpacing
                 do {
                     val lastTick = tickCount * tickSpacing
-                    if (lastTick in currentRange) {
+                    if (lastTick in currentRange.value) {
                         add(lastTick)
                     }
                     tickCount++
-                } while (lastTick < currentRange.last)
+                } while (lastTick < currentRange.value.last)
             }
         }
     }
@@ -115,7 +118,7 @@ public class IntLinearAxisModel(
         require(minTickSpacing > 0 && minTickSpacing <= 1) {
             "Minimum tick spacing must be greater than 0 and less than or equal to 1"
         }
-        val length = currentRange.last - currentRange.first
+        val length = currentRange.value.last - currentRange.value.first
         val magnitude = 10f.pow(floor(log10(length.toFloat())))
         val scaledTickRatios = TickRatios.map { (it * magnitude).roundToInt() }
 
@@ -147,21 +150,21 @@ public class IntLinearAxisModel(
             var i = 1
             do {
                 val nextTick = majorTickValues.last() + minorIncrement * i
-                if (nextTick in currentRange) {
+                if (nextTick in currentRange.value) {
                     add(nextTick)
                 }
                 i++
-            } while (nextTick in currentRange)
+            } while (nextTick in currentRange.value)
 
             // create ticks before first major tick. if still space in the range
             i = 1
             do {
                 val nextTick = majorTickValues.first() - minorIncrement * i
-                if (nextTick in currentRange) {
+                if (nextTick in currentRange.value) {
                     add(nextTick)
                 }
                 i++
-            } while (nextTick in currentRange)
+            } while (nextTick in currentRange.value)
         }
     }
 
@@ -171,16 +174,17 @@ public class IntLinearAxisModel(
         require(zoomFactor > 0) { "Zoom amount must be greater than 0" }
         require(pivot in 0.0..1.0) { "Zoom pivot must be between 0 and 1: $pivot" }
 
-        if (zoomFactor > 1f && currentRange.last - currentRange.first == minViewExtent) {
+        if (zoomFactor > 1f && currentRange.value.last - currentRange.value.first == minViewExtent) {
             // Can't zoom in more
-        } else if (zoomFactor < 1f && currentRange.last - currentRange.first == maxViewExtent) {
+        } else if (zoomFactor < 1f && currentRange.value.last - currentRange.value.first == maxViewExtent) {
             // Can't zoom out more
         } else {
             // convert pivot to axis range space
-            val pivotAxisScale = currentRange.first + ((currentRange.last - currentRange.first) * pivot).roundToInt()
+            val pivotAxisScale =
+                currentRange.value.first + ((currentRange.value.last - currentRange.value.first) * pivot).roundToInt()
 
-            val newLow = (pivotAxisScale - (pivotAxisScale - currentRange.first) / zoomFactor).roundToInt()
-            val newHi = (pivotAxisScale + (currentRange.last - pivotAxisScale) / zoomFactor).roundToInt()
+            val newLow = (pivotAxisScale - (pivotAxisScale - currentRange.value.first) / zoomFactor).roundToInt()
+            val newHi = (pivotAxisScale + (currentRange.value.last - pivotAxisScale) / zoomFactor).roundToInt()
 
             setViewRange(newLow..newHi)
         }
@@ -190,16 +194,16 @@ public class IntLinearAxisModel(
         if (!allowPanning) return
 
         // convert pan amount to axis range space
-        val panAxisScale = ((currentRange.last - currentRange.first) * amount).roundToInt()
+        val panAxisScale = ((currentRange.value.last - currentRange.value.first) * amount).roundToInt()
 
         // Limit pan amount to not exceed bounds of range
-        val panLimitEnd = min(panAxisScale, range.last - currentRange.last)
-        val panLimited = max(panLimitEnd, range.first - currentRange.first)
+        val panLimitEnd = min(panAxisScale, range.last - currentRange.value.last)
+        val panLimited = max(panLimitEnd, range.first - currentRange.value.first)
 
-        val newLow = (currentRange.first + panLimited)
-        val newHi = (currentRange.last + panLimited)
+        val newLow = (currentRange.value.first + panLimited)
+        val newHi = (currentRange.value.last + panLimited)
 
-        currentRange = newLow..newHi
+        currentRange.value = newLow..newHi
     }
 
     override fun setViewRange(newRange: ClosedRange<Int>) {
@@ -208,22 +212,22 @@ public class IntLinearAxisModel(
 
         if (newHi - newLow < minViewExtent) {
             val delta = (minViewExtent - (newHi - newLow)) / 2
-            currentRange = (newLow - delta)..(newHi + delta)
-            if (currentRange.first < range.first) {
-                currentRange = range.first..(range.first + minViewExtent)
-            } else if (currentRange.last > range.last) {
-                currentRange = (range.last - minViewExtent)..range.last
+            currentRange.value = (newLow - delta)..(newHi + delta)
+            if (currentRange.value.first < range.first) {
+                currentRange.value = range.first..(range.first + minViewExtent)
+            } else if (currentRange.value.last > range.last) {
+                currentRange.value = (range.last - minViewExtent)..range.last
             }
         } else if (newHi - newLow > maxViewExtent) {
             val delta = (newHi - newLow - maxViewExtent) / 2
-            currentRange = (newLow + delta)..(newHi - delta)
-            if (currentRange.first < range.first) {
-                currentRange = range.first..(range.first + maxViewExtent)
-            } else if (currentRange.last > range.last) {
-                currentRange = (range.last - maxViewExtent)..range.last
+            currentRange.value = (newLow + delta)..(newHi - delta)
+            if (currentRange.value.first < range.first) {
+                currentRange.value = range.first..(range.first + maxViewExtent)
+            } else if (currentRange.value.last > range.last) {
+                currentRange.value = (range.last - maxViewExtent)..range.last
             }
         } else {
-            currentRange = newLow..newHi
+            currentRange.value = newLow..newHi
         }
     }